Cast array accessors from char to int to make gcc shut up.

+++ kmer/libbio/reversecomplement.c
@@ -12,13 +12,13 @@ reverseComplementSequence(char *seq, u32bit seqlen) {
   u32bit  c = seqlen / 2;
 
   while (c--) {
-    t = complementSymbol[*s];
-    *(s++) = complementSymbol[*e];
+    t = complementSymbol[(int)*s];
+    *(s++) = complementSymbol[(int)*e];
     *(e--) = t;
   }
 
   if (s == e)
-    *s = complementSymbol[*s];
+    *s = complementSymbol[(int)*s];
 
   return(seq);
 }
